Rename electrostatic_potential.py to esp.py
Change the code, docstrings, notes and file names to reflect this renaming.
