+++ docs/notes.tex
@@ -1380,7 +1380,7 @@ matrix is given by \verb|one_dm|.
 \item To compute the electrostatic potential using density matrix expressed with
   respect to atomic orbitals
   \begin{lstlisting}[xleftmargin=-25pt]
-from gbasis.evals.electrostatic_potential import electrostatic_potential
+from gbasis.evals.esp import electrostatic_potential
 
 output = electrostatic_potential(
     basis,
@@ -1394,7 +1394,7 @@ output = electrostatic_potential(
 \item To compute the electrostatic potential using density matrix expressed with
   respect to molecular orbitals
   \begin{lstlisting}[xleftmargin=-25pt]
-from gbasis.evals.electrostatic_potential import electrostatic_potential
+from gbasis.evals.esp import electrostatic_potential
 
 output = electrostatic_potential(
     basis,
